create console_widget folder
