+++ src/admc/console_widget/console_widget.h
