+++ src/admc/console_widget/console_widget_p.h
@@ -24,7 +24,7 @@
  * Private header for ConsoleWidget.
  */
 
-#include "results_view.h"
+#include "console_widget/results_view.h"
 
 class QLabel;
 class QStackedWidget;
