+++ src/admc/console_widget/results_description.h
