+++ src/admc/console_widget/results_view.h
